+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 -   alt modifier can be used to disable draw/resize snapping
 -   Client option to invert the ALT behaviour (i.e. invert snapping behaviour)
 -   Logo, version info and some urls to the topleft section which is only visible if both locations bar and menu are opened
+-   Search bar to asset menu
 
 ### Fixed